refactor: revert some changes from previous commits.
