+++ tests/integration/streamlit/test_video_renderer.py
@@ -1,12 +1,12 @@
-from collections.abc import Callable
 from pathlib import Path
 
 import cv2
 from starlette.testclient import TestClient
 
 from vigil.adapters.primary.streamlit.components.api_client import VigilClient
-from vigil.adapters.primary.streamlit.components.models import DetectionData, TrackData
 from vigil.adapters.primary.streamlit.components.video_renderer import render_video_with_tracks
+
+from tests.integration.streamlit.conftest import make_detection, make_track
 
 
 def test_renders_a_readable_mp4_file(video_path: Path) -> None:
@@ -31,11 +31,7 @@ def test_output_has_same_frame_count_as_input(video_path: Path) -> None:
     assert frame_count == 10
 
 
-def test_renders_without_error_when_detections_span_all_frames(
-    video_path: Path,
-    make_detection: Callable[..., DetectionData],
-    make_track: Callable[..., TrackData],
-) -> None:
+def test_renders_without_error_when_detections_span_all_frames(video_path: Path) -> None:
     tracks = [
         make_track(
             detections=tuple(make_detection(frame_position=i) for i in range(10)),
