+++ vigil/adapters/primary/fastapi/controllers/video_tracks_retrieval.py
@@ -1,15 +1,13 @@
 from dataclasses import dataclass
 from uuid import UUID
 
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends
 from pydantic import BaseModel
 
 from vigil.adapters.primary.fastapi.app_dependencies import (
     _get_frame_repository,
-    _get_video_repository,
     get_video_tracks_use_case,
 )
-from vigil.business_logic.exceptions import VideoNotFoundError
 
 router = APIRouter(tags=["videos"])
 
@@ -61,14 +59,8 @@ def get_video_tracks(
     video_id: UUID,
     use_case=Depends(get_video_tracks_use_case),
     frame_repository=Depends(_get_frame_repository),
-    video_repository=Depends(_get_video_repository),
 ) -> GetVideoTracksResponse:
     """Return all tracks (open and closed) for the given video."""
-    try:
-        video_repository.frame_count(video_id)
-    except VideoNotFoundError as err:
-        raise HTTPException(status_code=404, detail="Video not found.") from err
-
     tracks = use_case.execute(video_id=video_id)
     return GetVideoTracksResponse(
         status="success",
